pyproject.toml
[project]
name = "llm-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "edge-tts>=7.2.7",
    "langchain>=1.2.9",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
    "trafilatura>=2.0.0",
]
